'creating_a_query_to_show_existing_bookings'
