--- People/deleting_passenger_info_from_database.py
@@ -1,4 +1,0 @@
-
-class Deleting_passenger_info():
-    def delete_info(self):
-        pass
