+++ database/adding_passenger_info_to_database.py
@@ -1,8 +1,7 @@
-import pyodbc
 import time
-from create_connection import Database_OOP
-from menu_interface import Menu_interface
-from welcome_interface import Welcome_interface
+from database.create_connection import Database_OOP
+from FrontEndInterface.menu_interface import Menu_interface
+from FrontEndInterface.welcome_interface import Welcome_interface
 
 """
 In this class we attempt to add the information inputted from the airport assistant to build a database
