+++ database/show_existing_booking.py
@@ -1,0 +1,11 @@
+import time
+from database.create_connection import Database_OOP
+from FrontEndInterface.menu_interface import Menu_interface
+from FrontEndInterface.welcome_interface import Welcome_interface
+
+class Show_existing_booking():
+
+    def existing_booking_query(self):
+        obj1 = Database_OOP()
+        cursor = obj1.establish_connection()
+        query = "SELECT * FROM Bookings WHERE  "
